Adding the main form
Added the main slides for the registration. Their are 4 components that are imported into the form depending on which slide the user is on. There are 2 buttons added to the bottom of the form which allows the user to progress through the slides and then submit on the final part of the form.